this should work... I hope...

+++ buildtools/deploy.sh
@@ -7,7 +7,14 @@ VERSION=$2
 
 pushd $DEPLOY_DIR
 
+file=$VERSIONS/NurseBot$VERSION.jar
+
+if test ! -f $file; then
+	echo "Version not found."
+	exit 1
+fi
+
 unlink NurseBot.jar
-ln $VERSIONS/NurseBot$VERSION.jar NurseBot.jar
+ln $file NurseBot.jar
 
 popd
